docs: youtube-music is not a userstyle
As discussed on discord, the youtube-music userstyle has
separate css files and supports desktop clients alongside
stylus. As this is now similar to the discord port
which we are currently not including, we should not label
this as a userstyle.

We may come back and revisit this decision at a later point
in time.