refactor(ports.yml): remove nickname from `collaborators`
I was finding that this field was inconsistent sometimes if the
collaborator was both in the `userstyles.yml` and `ports.yml`. Just
simplifying it to one `url` field for now.

At some point in the future, might get rid of the `url` field entirely
and instead rely on the GitHub API, but I'd rather not introduce a
build time for retrieving the names of the maintainers right now.